Split functionality accross multiple modules
