+++ gui/src/main.rs
@@ -1,9 +1,11 @@
-use std::path::Path;
-
 use recursive_hash_calculator_core::{hash_strategy::HashStrategy, hasher, report_type::ReportType};
-use rfd::FileDialog;
+use report_path::report_output_path;
 
 slint::include_modules!();
+
+mod file_dialog;
+mod report_path;
+mod result_formatter;
 
 #[tokio::main]
 async fn main() -> Result<(), slint::PlatformError> {
@@ -11,12 +13,12 @@ async fn main() -> Result<(), slint::PlatformError> {
 
     window.on_input_path_button_clicked({
         let window = window.as_weak().unwrap();
-        move || window.set_input_path(open_directory_input().into())
+        move || window.set_input_path(file_dialog::open_directory_dialog().into())
     });
 
     window.on_output_path_button_clicked({
         let window = window.as_weak().unwrap();
-        move || window.set_output_path(open_directory_input().into())
+        move || window.set_output_path(file_dialog::open_directory_dialog().into())
     });
 
     window.on_calculate_button_clicked({
@@ -42,48 +44,25 @@ async fn main() -> Result<(), slint::PlatformError> {
 
             window.set_buttons_enabled(false);
 
-            let output_path = Path::new(&output_path)
-                .join(report_filename())
-                .to_str()
-                .unwrap()
-                .to_owned();
-
-            tokio::spawn(calculate_hashes(weak_window.clone(), input_path, output_path));
+            tokio::spawn(calculate_hashes(
+                weak_window.clone(),
+                input_path,
+                report_output_path(&output_path),
+            ));
         }
     });
 
     window.run()
 }
 
-fn open_directory_input() -> String {
-    let directory = FileDialog::new().pick_folder();
-
-    match directory {
-        Some(path) => path.to_str().unwrap_or_default().to_owned(),
-        None => String::default(),
-    }
-}
-
 async fn calculate_hashes(window: slint::Weak<MainWindow>, input_path: String, output_path: String) {
     let result = hasher::process(input_path, HashStrategy::Sha256, ReportType::Csv(output_path.clone())).await;
+    let info_message = result_formatter::format(&result, &output_path);
 
     window
         .upgrade_in_event_loop(move |window| {
-            let count = result.processed_files_count;
-            let elapsed_time = result.elapsed_time.as_secs_f32();
-            let files_have = if count == 1 { "file has" } else { "files have" };
-
             window.set_buttons_enabled(true);
-            window.set_info_message(
-                format!(
-                    "{count} {files_have} been processed!\n\nTook {elapsed_time} seconds.\n\nThe report has been saved at:\n{output_path}",
-                )
-                .into(),
-            );
+            window.set_info_message(info_message.into());
         })
         .unwrap();
-}
-
-fn report_filename() -> String {
-    chrono::Local::now().format("%Y-%m-%d_%H%M%S.csv").to_string()
 }
